refactor(api): extract collections router (pure move)
Move the 5 collection routes (list/create/get/update/delete) plus the
domain-local _validate_collection_payload helper from api_server.py to
src/collections/routes.py, included at the block's old position.
Handler bodies are verbatim; only imports were adjusted.

get_collection_service / get_page_service / get_template_engine resolve
through src.api_server at call time — the suite patches them there.
CollectionCreate/CollectionUpdate imports move with the handlers;
is_collection_id stays imported in api_server (still used by the
active-page resolvers).

Response-shape goldens: zero changes. Route-inventory golden: zero
changes.

Part of #1849; closes #1756.
